requirements.txt
django
Authlib
requests
charset-normalizer==3.1.0
python-dotenv
gunicorn
djangorestframework